Add another optimization using trie structure
Using trie to search a string is really fast. The complexity of
searching a string is O(n), with n equals to the length of the
string. However, this method needs a large amount of space to
store data.

In this program, I assume lastName contains only upper/lower
case alphbets. If the string has other characters, append()
won't place the string into memory, and similar things will
do to findName(), too.

Compared to hash function, this method has lower performance
in append() and also need more memory space, and merely the same
performance in findName(). However, it's good practice to try
using a new method to solve the problem.